+++ docs/changelog.rst
@@ -1,6 +1,10 @@
 Changelog
 ---------
 
+* `Next Release`_
+
+  - Modify content negotiation to 406 when asked for an unknown character set
+
 * `0.0.3`_ (30 May 2015)
 
   - Add :func:`glinda.content.clear_handlers`